refactor(control-plane): split stale-recovery integration suite, honest threshold docs
