+++ packages/control-plane/src/image-builds/maintenance.ts
@@ -6,7 +6,7 @@ const SNAPSHOT_FILESYSTEM_TIMEOUT_SECONDS = 300;
 const BUILD_FUNCTION_TIMEOUT_MARGIN_SECONDS = 300;
 
 /**
- * Age past which a `building` row is provably dead and safe to fail: sized at
+ * Age past which a `building` row is presumed dead and safe to fail: sized at
  * the longest possible build's worker lifetime so a long-but-live build is
  * never reaped. Modal is the worst case — sandbox lifetime is capped at
  * MAX_BUILD_TIMEOUT_SECONDS, and its build worker idles through the snapshot
@@ -14,8 +14,16 @@ const BUILD_FUNCTION_TIMEOUT_MARGIN_SECONDS = 300;
  * packages/modal-infra/src/sandbox/manager.py). Vercel and OpenComputer
  * ceilings sit at or under this; maintenance.test.ts pins all three, so a
  * provider lifetime bump past the threshold fails CI instead of live-failing
- * long builds. The clock starts at row registration (`created_at`), not
- * sandbox start, so provider queueing eats into the margin.
+ * long builds.
+ *
+ * The clock starts at row registration (`created_at`), not sandbox start, so
+ * dispatch latency and provider queueing eat into the margin: a build that
+ * both runs to its full timeout AND is queued longer than the margin before
+ * executing would be misclassified. That residual window is inherited from
+ * the cron sweep (the Modal scheduler has always posted this same threshold
+ * to /image-builds/mark-stale), and the blast radius is a redundant rebuild —
+ * the misclassified build's late callback is rejected by the
+ * `status = 'building'` completion guards, never recorded over the new build.
  */
 export const DEFAULT_STALE_BUILD_MAX_AGE_MS =
   (MAX_BUILD_TIMEOUT_SECONDS +
